[ * ] index, music, movie .jsp 컨텐츠 추가 및 삽입 검색창 기능 구현, boardmain.jsp 수정 및 보완 로그인&회원가입 기능구현 및 디자인 완성

+++ loginbox.jsp
@@ -26,16 +26,31 @@
 }
 
 </style>
-
+<%
+    // 세션에서 사용자 정보 가져오기
+    String user_id = (String) session.getAttribute("userID");
+    String user_name = (String) session.getAttribute("userNAME"); 
+%>
 <div class="section login-box">
-	<h4>유벤에 로그인 하세요</h4>
-	<form action="loginAction.jsp" method="post">
-		<input type="text" name="userid" placeholder="아이디"> <input
-			type="password" name="password" placeholder="비밀번호">
-		<button type="submit" name="login">로그인</button>
-		<button type="button" onclick="location.href='register.jsp'">회원가입</button>
-		<button type="button" onclick="location.href='delete.jsp'">회원탈퇴</button>
-		<button type="button" onclick="location.href='profile.jsp'">내
-			정보</button>
-	</form>
+	<% if (user_id != null) { %>
+        <h4><%= user_name %>님 환영합니다!</h4>
+        
+        <div class="logged-in-buttons">
+            <button type="button" onclick="location.href='logout.jsp'">로그아웃</button>
+            <button type="button" onclick="location.href='delete.jsp'">회원탈퇴</button>
+            <button type="button" onclick="location.href='profile.jsp'">내정보</button>
+        </div>
+        
+    <% } else { %>
+        <h4>유벤에 로그인 하세요</h4>
+        <form action="loginAction.jsp" method="post">
+            <input type="text" name="userid" placeholder="아이디"> 
+            <input type="password" name="password" placeholder="비밀번호">
+            <button type="submit" name="login">로그인</button>
+            <button type="button" onclick="location.href='register.jsp'">회원가입</button>
+            
+            <button type="button" onclick="location.href='delete.jsp'">회원탈퇴</button>
+            <button type="button" onclick="location.href='profile.jsp'">내정보</button>
+        </form>
+    <% } %>
 </div>
